course attributes dictionary now governs all its data
courses will contain a dictionary named attributes. It stores attributes in the form of A.B.C.....Z (period delimited strings).

We can then match attributes to target values to determine its properties and filter courses.

The point of this steaming pile of buggy code is to allow courses to be more generalized and does not rely on predefined attributes to function.