No replacement course selection via priority selection array
If we encounter a rule that does not allow replacement, we attempt to seek a combination of course selections that minimally affects the excess fulfillment of previous courses and absolutely does not affect the fulfillment of previous courses.

Excess fulfillment refers to a previous courses's fulfillment set being larger than the required count. Fulfillment refers to any course which does not have excess fulfillment courses.

The algorithm used requires the input rules in order of importance, where earlier rules will be given priority at filling its requirements. This guarantees that later rules will not negatively impact the fulfillment of an earlier rule.

Alternatively, a rule can be labeled as 'high priority' to automatically put it at the front of the queue, before any rule not labelled as such.