added visualization support
Output will contain a static dictionary containing all visualization objects. Visualization objects can by user defined, so long they contain a method "visualize()" with any amount of parameters.

To add a visualizer, simply add it to Output.visualizers with the name as key.

To use a visualization, call Output.visualize(visualization_name, *args), where visualization_name is the name you used when you added it earlier, and *args are the arguments you would enter into the visualizer.

This method allows you to write visualization output anywhere within your code, and have it execute the appropriate visualization object, or have it not execute at all by simply not adding it to the visualizers dictionary.